requirements.txt
pyserial
flask
dash
dash-html-components
dash-bootstrap-components
dash-leaflet
dash-core-components
dash-daq
plotly
xlwt
open-cv
requests
datetime
